+++ lib/apiVersion.js
@@ -1,3 +1,5 @@
 // File generated from our OpenAPI spec
 
-module.exports = {ApiVersion: '2020-08-27; server_side_confirmation_beta=v1'};
+module.exports = {
+  ApiVersion: '2022-08-01; server_side_confirmation_beta=v1; orders_beta=v4',
+};
